+++ src/Components/DetailPage/Detail.tsx
