+++ src/app/community/page.tsx
@@ -1,4 +1,5 @@
 import Filter from "@/Components/Community/Filter";
+
 import React from "react";
 
 const ComuunityPage = () => {
